pyproject.toml
[project]
name = "fastrtc-voice-agent"
version = "0.1.4"
description = "A modular voice agent with swappable STT/TTS/LLM backends"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Clement Poisson" }]
keywords = ["voice", "agent", "stt", "tts", "llm", "whisper", "edge-tts", "fastrtc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "numpy>=1.24",
    "librosa>=0.10",
    "edge-tts>=7.2.7",
    "pydub>=0.25.1",
    "faster-whisper>=1.2.1",
    "ollama>=0.6.1",
    "fastrtc[vad]>=0.0.34",
    "pip>=25.3",
    "twine>=6.2.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "dotenv>=0.9.9",
    "groq>=1.0.0",
]

[project.optional-dependencies]
whisper = ["openai-whisper>=20250625"]
groq = ["groq>=0.11.0"]
ollama = ["ollama>=0.6.1"]
claude = ["anthropic>=0.40.0", "python-dotenv>=1.0.0"]
all = [
    "fastrtc-voice-agent[whisper,groq,ollama,claude]",
]

[project.scripts]
fastrtc-voice-agent = "voice_agent.cli:main"

[project.urls]
Repository = "https://github.com/SonDePoisson/voice-agent"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_agent"]

[tool.hatch.build.targets.sdist]
include = ["src/voice_agent"]
